# perf(s2n-quic-core): optimize checksums for small payloads